Use auto-generated in.json files
Use a program to generate the expected in.json files automatically from the test.event files.

Fixes issue #4.

Note that test 2JQS is deleted because the YAML appears to be invalid.

Also, it's not currently possible to have in.json files for multi-document YAML documents, as there's only a single JSON object in each file. That could be solved either by always providing an array of JSON objects, or by providing a sequence of JSON objects.